Add newline before key in authorized_keys
This script appends a pub ssh key from djeep to authorized keys, however
if authorized keys didn't have a newline at the end, the key would get
  appended to the same line as the previous key and not work.

This patch adds a newline before the key, so that the append will
definitely work regardless of whether authorized_keys had a newline at
EOF or not.

Also overwrite pub key rather than appending to it.